feat(about): move role subcommand from /role info (#273)
